# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Invite {
  token: string;
  groupId: number;
  groupName: string;
  createdAt: string;
}

export interface InvitePreview {
  token: string;
  groupName: string;
  inviterName: string;
  valid: boolean;
}

export interface AcceptInviteResult {
  joined: boolean;
  /** @nullable */
  groupId: number | null;
  groupName: string;
}

export interface GroupInput {
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
}

export type GroupRole = typeof GroupRole[keyof typeof GroupRole];


export const GroupRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface Group {
  id: number;
  name: string;
  ownerId: string;
  role: GroupRole;
  memberCount: number;
  createdAt: string;
}

export type GroupMemberRole = typeof GroupMemberRole[keyof typeof GroupMemberRole];


export const GroupMemberRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface GroupMember {
  userId: string;
  displayName: string;
  role: GroupMemberRole;
  joinedAt: string;
}

export type GroupDetailRole = typeof GroupDetailRole[keyof typeof GroupDetailRole];


export const GroupDetailRole = {
  owner: 'owner',
  member: 'member',
} as const;

export interface GroupDetail {
  id: number;
  name: string;
  ownerId: string;
  role: GroupDetailRole;
  memberCount: number;
  createdAt: string;
  members: GroupMember[];
}

export type MediaType = typeof MediaType[keyof typeof MediaType];


export const MediaType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface Entry {
  id: number;
  title: string;
  mediaType: MediaType;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  category: string;
  /** @nullable */
  comment?: string | null;
  /** Display name of the user who added this entry. */
  addedBy: string;
  createdAt: string;
}

export interface EntryInput {
  /** @minLength 1 */
  title: string;
  mediaType: MediaType;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 1 */
  category: string;
  comment?: string;
}

export interface EntryUpdate {
  /** @minLength 1 */
  title?: string;
  mediaType?: MediaType;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  /** @minLength 1 */
  category?: string;
  /** @nullable */
  comment?: string | null;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface Stats {
  total: number;
  movieCount: number;
  tvCount: number;
  averageRating: number;
  byCategory: CategoryCount[];
}

export interface WatchlistItem {
  id: number;
  title: string;
  mediaType: MediaType;
  createdAt: string;
  /** Display names of other group members who rated or saved this show. */
  alsoEngagedBy: string[];
}

export interface WatchlistInput {
  /** @minLength 1 */
  title: string;
  mediaType: MediaType;
}

export type Approval = typeof Approval[keyof typeof Approval];


export const Approval = {
  yes: 'yes',
  no: 'no',
  solo: 'solo',
} as const;

export interface ShowApproval {
  titleKey: string;
  mediaType: MediaType;
  yes: number;
  no: number;
  solo: number;
  /** The current user's own answer, or null if they haven't answered. */
  myApproval: Approval | null;
}

export interface ApprovalInput {
  /** @minLength 1 */
  title: string;
  mediaType: MediaType;
  approval: Approval;
  /** Scope the returned tallies to members of this group. Without it, only the caller's own answer is counted. */
  groupId?: number;
}

export interface CheckInResult {
  /** Entries added by other group members since the previous visit. */
  newCount: number;
  /**
     * The previous last-seen time, or null on first visit.
     * @nullable
     */
  since: string | null;
}

export type ListEntriesParams = {
/**
 * View this member's entries instead of the group's. Requires a shared group.
 */
userId?: string;
/**
 * Scope the list to members of this group. You must be a member.
 */
groupId?: number;
category?: string;
mediaType?: MediaType;
sort?: ListEntriesSort;
};

export type ListEntriesSort = typeof ListEntriesSort[keyof typeof ListEntriesSort];


export const ListEntriesSort = {
  newest: 'newest',
  oldest: 'oldest',
  rating_high: 'rating_high',
  rating_low: 'rating_low',
  title: 'title',
} as const;

export type GetStatsParams = {
/**
 * View this member's stats instead of the group's. Requires a shared group.
 */
userId?: string;
/**
 * Aggregate stats across members of this group. You must be a member.
 */
groupId?: number;
};

export type ListWatchlistParams = {
/**
 * View this member's saved shows instead of your own. Requires a shared group.
 */
userId?: string;
/**
 * Scope the "also engaged by" names to members of this group.
 */
groupId?: number;
};

export type ListApprovalsParams = {
/**
 * Scope the tallies to members of this group.
 */
groupId?: number;
};

export type ClearApprovalParams = {
title: string;
mediaType: MediaType;
/**
 * Scope the returned tallies to members of this group.
 */
groupId?: number;
};

export type CheckInParams = {
/**
 * Count new entries from other members of this group.
 */
groupId?: number;
};