# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List watchlist entries
      description: Returns entries from members of the active group by default. Pass groupId to scope to a specific group (you must be a member), or userId to view a single member's entries (only allowed when you share a group). With neither, returns entries from everyone who shares a group with you.
      parameters:
        - name: userId
          in: query
          required: false
          description: View this member's entries instead of the group's. Requires a shared group.
          schema:
            type: string
        - name: groupId
          in: query
          required: false
          description: Scope the list to members of this group. You must be a member.
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: mediaType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/MediaType"
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [newest, oldest, rating_high, rating_low, title]
      responses:
        "200":
          description: A list of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a watchlist entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a single entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /stats:
    get:
      operationId: getStats
      tags: [entries]
      summary: Watchlist summary stats
      description: Aggregate counts, average rating, and category breakdown. Pass groupId to aggregate across all members of a group (you must be a member), or userId for a single member's stats (only allowed when you share a group). With neither, aggregates across everyone who shares a group with you.
      parameters:
        - name: userId
          in: query
          required: false
          description: View this member's stats instead of the group's. Requires a shared group.
          schema:
            type: string
        - name: groupId
          in: query
          required: false
          description: Aggregate stats across members of this group. You must be a member.
          schema:
            type: integer
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
...
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List "Wife Approved?" summaries
      description: >-
        Returns, for every show that has at least one answer within the given
        group, the count of each answer across that group's members plus the
        caller's own answer. Without a groupId, only the caller's own answers
        are counted.
      parameters:
        - name: groupId
          in: query
          required: false
          description: Scope the tallies to members of this group.
          schema:
            type: integer
      responses:
        "200":
          description: Approval summaries by show
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShowApproval"
    put:
      operationId: setApproval
      tags: [approvals]
      summary: Set my "Wife Approved?" answer for a show
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated approval summary for the show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowApproval"
    delete:
      operationId: clearApproval
      tags: [approvals]
      summary: Clear my answer for a show
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: mediaType
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/MediaType"
        - name: groupId
          in: query
          required: false
          description: Scope the returned tallies to members of this group.
          schema:
            type: integer
      responses:
        "200":
          description: Updated approval summary for the show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowApproval"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MediaType:
      type: string
      enum: [movie, tv]
    Entry:
      type: object
      required:
        - id
        - title
        - mediaType
        - rating
        - category
        - addedBy
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        mediaType:
          $ref: "#/components/schemas/MediaType"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        category:
          type: string
        comment:
          type: ["string", "null"]
        addedBy:
          type: string
          description: Display name of the user who added this entry.
        createdAt:
          type: string
          format: date-time
    EntryInput:
      type: object
      required:
        - title
        - mediaType
        - rating
        - category
      properties:
        title:
          type: string
          minLength: 1
        mediaType:
          $ref: "#/components/schemas/MediaType"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        category:
          type: string
          minLength: 1
        comment:
          type: string
    EntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        mediaType:
          $ref: "#/components/schemas/MediaType"
        rating:
          type: integer
          minimum: 1
          maximum: 5
        category:
          type: string
          minLength: 1
        comment:
          type: ["string", "null"]
    CategoryCount:
      type: object
      required:
        - category
        - count
      properties:
        category:
          type: string
        count:
          type: integer
    Stats:
      type: object
      required:
        - total
        - movieCount
        - tvCount
        - averageRating
        - byCategory
      properties:
        total:
          type: integer
        movieCount:
          type: integer
        tvCount:
          type: integer
        averageRating:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
...
    Approval:
      type: string
      enum: [yes, no, solo]
    ShowApproval:
      type: object
      required:
        - titleKey
        - mediaType
        - "yes"
        - "no"
        - solo
        - myApproval
      properties:
        titleKey:
          type: string
        mediaType:
          $ref: "#/components/schemas/MediaType"
        "yes":
          type: integer
        "no":
          type: integer
        solo:
          type: integer
        myApproval:
          description: The current user's own answer, or null if they haven't answered.
          anyOf:
            - $ref: "#/components/schemas/Approval"
            - type: "null"
    ApprovalInput:
      type: object
      required:
        - title
        - mediaType
        - approval
      properties:
        title:
          type: string
          minLength: 1
        mediaType:
          $ref: "#/components/schemas/MediaType"
        approval:
          $ref: "#/components/schemas/Approval"
        groupId:
          type: integer
          description: >-
            Scope the returned tallies to members of this group. Without it,
            only the caller's own answer is counted.
...